esp_tg_bot/requirements.txt
aiogram==3.4.1
httpx==0.27.0
python-dotenv
debugpy